fix null derefence exception when surface is reset
Summary: when media player is stopped flush is called to flush the remaining buffer and the output surface is set to null.
However while the buffer is getting flushed the output can still be used causing the null dereference crash.
This diff adds a check in renderOneBufferItem_l to check if output is valid and go through the regular abandon flow instead of using the null output.

Test: successful build on master

Change-Id: Ic3d6707e2f8e0943ca2368a8e50ea683b6dac46a
Signed-off-by: Abdelrahman Daim <adaim@meta.com>
Signed-off-by: Ghosuto <clash.raja10@gmail.com>