Null Pointer Exception in CameraServer during user switch
A Null Pointer Exception (NPE) occurs in
Camera3Device::setNotifyCallback when switching users. The crash happens
upon accessing mRequestThread, which has already been destroyed.

The root cause is a race condition: the UidPolicy callback, triggered by
the user switch, disconnects the device and nullifies mRequestThread. If
setNotifyCallback is called after this, it leads to the NPE. The issue
is 100% reproducible by adding a delay in setNotifyCallback to simulate
the race.

Bug:457793937

Change-Id: I6d0bd5488ae773c4db1e87f7dc4813fe4d6636ab
Signed-off-by: Phoenix wang <wxymails@gmail.com>
Signed-off-by: Ghosuto <clash.raja10@gmail.com>